+++ sdk/src/androidTest/java/ly/count/android/sdk/CountlyTests.java
@@ -37,6 +37,7 @@ import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertSame;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
+import static org.mockito.ArgumentMatchers.isNull;
 import static org.mockito.Mockito.anyInt;
 import static org.mockito.Mockito.anyString;
 import static org.mockito.Mockito.mock;
@@ -401,7 +402,7 @@ public class CountlyTests {
         assertEquals(0, mCountly.getActivityCount());
         assertTrue(mCountly.getPrevSessionDurationStartTime() > 0);
         verify(requestQueueProvider).endSession(0);
-        verify(requestQueueProvider, times(1)).recordEvents(anyString()); // not 0 anymore, it will send orientation event
+        verify(requestQueueProvider, times(1)).recordEvents(anyString(), isNull()); // not 0 anymore, it will send orientation event
     }
 
     /**
@@ -428,7 +429,7 @@ public class CountlyTests {
         assertEquals(0, mCountly.getActivityCount());
         assertTrue(mCountly.getPrevSessionDurationStartTime() > 0);
         verify(requestQueueProvider).endSession(0);
-        verify(requestQueueProvider).recordEvents(eventStr);
+        verify(requestQueueProvider).recordEvents(eventStr, null);
     }
 
     @Test
@@ -493,7 +494,7 @@ public class CountlyTests {
         mCountly.moduleRequestQueue.sendEventsIfNeeded(false);
 
         verify(mCountly.config_.storageProvider, times(1)).getEventsForRequestAndEmptyEventQueue();
-        verify(requestQueueProvider, times(1)).recordEvents(eventData);
+        verify(requestQueueProvider, times(1)).recordEvents(eventData, null);
     }
 
     @Test
@@ -509,7 +510,7 @@ public class CountlyTests {
         mCountly.moduleRequestQueue.sendEventsIfNeeded(false);
 
         verify(mCountly.config_.storageProvider, times(1)).getEventsForRequestAndEmptyEventQueue();
-        verify(requestQueueProvider, times(1)).recordEvents(eventData);
+        verify(requestQueueProvider, times(1)).recordEvents(eventData, null);
     }
 
     @Test
@@ -534,7 +535,7 @@ public class CountlyTests {
         mCountly.onTimer();
 
         verify(requestQueueProvider).updateSession(0);
-        verify(requestQueueProvider, times(1)).recordEvents(anyString()); // not 0 anymore, it will send orientation event
+        verify(requestQueueProvider, times(1)).recordEvents(anyString(), isNull()); // not 0 anymore, it will send orientation event
     }
 
     @Test
@@ -551,7 +552,7 @@ public class CountlyTests {
         mCountly.onTimer();
 
         verify(requestQueueProvider).updateSession(0);
-        verify(requestQueueProvider).recordEvents(eventData);
+        verify(requestQueueProvider).recordEvents(eventData, null);
     }
 
     @Test