+++ sdk/src/main/java/ly/count/android/sdk/ConfigurationProvider.java
@@ -1,5 +1,8 @@
 package ly.count.android.sdk;
 
+import java.util.Map;
+import java.util.Set;
+
 interface ConfigurationProvider {
     boolean getNetworkingEnabled();
 
@@ -31,4 +34,28 @@ interface ConfigurationProvider {
     int getBOMDuration();
 
     int getRequestTimeoutDurationMillis();
+
+    int getUserPropertyCacheLimit();
+
+    // LISTING FILTERS
+
+    FilterList<Set<String>> getEventFilterList();
+
+    FilterList<Set<String>> getUserPropertyFilterList();
+
+    FilterList<Set<String>> getSegmentationFilterList();
+
+    FilterList<Map<String, Set<String>>> getEventSegmentationFilterList();
+
+    Set<String> getJourneyTriggerEvents();
+
+    class FilterList<T> {
+        T filterList;
+        boolean isWhitelist;
+
+        FilterList(T filterList, boolean isWhitelist) {
+            this.filterList = filterList;
+            this.isWhitelist = isWhitelist;
+        }
+    }
 }