+++ sdk/src/main/java/ly/count/android/sdk/CountlyConfig.java
@@ -213,6 +213,12 @@ public class CountlyConfig {
     // If set to true, request queue cleaner will remove all overflow at once instead of gradually (loop limited) removing
     boolean disableGradualRequestCleaner = false;
 
+    // If set to true, the SDK will not store the default push consent state on initialization for not requiring consent
+    boolean disableStoringDefaultPushConsent = false;
+
+    // If set to true, the SDK will not restart manual views while switching between foreground and background
+    boolean disableViewRestartForManualRecording = false;
+
     /**
      * THIS VARIABLE SHOULD NOT BE USED
      * IT IS ONLY FOR INTERNAL TESTING
@@ -1063,7 +1069,7 @@ public class CountlyConfig {
      * Set the webview display option for Content and Feedback Widgets
      *
      * @param displayOption IMMERSIVE for full screen with hidden system UI, or
-     *                      SAFE_AREA to use app usable area and not overlap system UI
+     * SAFE_AREA to use app usable area and not overlap system UI
      * @return config content to chain calls
      */
     public synchronized CountlyConfig setWebviewDisplayOption(WebViewDisplayOption displayOption) {
@@ -1097,6 +1103,32 @@ public class CountlyConfig {
      */
     public synchronized CountlyConfig disableGradualRequestCleaner() {
         this.disableGradualRequestCleaner = true;
+        return this;
+    }
+
+    /**
+     * Disable storing the default push consent on initialization.
+     * By default, if consent is required and push consent is not set,
+     * the SDK was storing push consent as false on initialization.
+     * Now, if consent is not required, the SDK will store push consent as true on initialization.
+     *
+     * @return Returns the same config object for convenient linking
+     */
+    public synchronized CountlyConfig disableStoringDefaultPushConsent() {
+        this.disableStoringDefaultPushConsent = true;
+        return this;
+    }
+
+    /**
+     * Disable view restart when manual view recording is done.
+     * By default, if automatic view tracking is not enabled and a manual view is recorded,
+     * the SDK was restarting the views to properly track the view duration in bg/fg transitions.
+     * Now, with this option enabled, the SDK will not restart the views on manual view recording.
+     *
+     * @return Returns the same config object for convenient linking
+     */
+    public synchronized CountlyConfig disableViewRestartForManualRecording() {
+        this.disableViewRestartForManualRecording = true;
         return this;
     }